Add TEXMFDBS configuration
We had somehow lost my per-user configuration, so per-user files were
not found.

Adding the per-user path to TEXINPUTS didn’t work, because apparently
kpathsea stops as soon as it hits a directory or file for which it lacks
permission, and some of my texmf hierarchy is private.

However, ls-R does work (provided it’s up to date!).

Leave a FIXME to configure rather than hardwire the /home/rrt path.