Make LaTeX to PDF converter work with stdin input
This involves two main changes: first, if the input is supplied on
stdin (file name is “-”) then write it to a file, as we can’t otherwise
process it. Secondly, extend rather than overwrite the TEXINPUTS and
BIBINPUTS environment variables, so that they can be supplied from the
environment. This is crucial for files referenced in the same directory as
the original source file, as we don’t know its path!